rename file base.ts to clientBase.ts
